Fixed how we refer to the grid matrix
There was an error found within the code. Here is a brief description what the error was:

The function make_grid() makes a matrix from lists and inserts the data in in a (x,y) format where x was the row index, and y was the column index. This is how we access and read from the matrix.

However, we find in the function getNode() which uses pygame to grab the coordinates on the screen of where our mouse clicked, and it returns the column and row we clicked. The issue with this is that pygame returns the mouse position based on a (y,x) format persay, where the column is the first value and the row is the second. So with this in mind, the function getNode() returns back a value format of (Column, Row). And this function is used throughout the code, and most the code was built off of a Column, Row system due to this function returning it in this format. This makes it so that if want want to ouput the grid in a terminal, or set up a correct moves list, we have to refer to everything in reverse due to how the grid is defined in a (row, column) format.

I went ahead and changed all the pieces of code where we refer to grabbing a piece from the matrix, in a (row, column) format, the same format we used for declaring the grid