Currently Working on checkForPins
Currently working on fully implementing checkForPins function to check all positions that could possibly pin the king, and preventing the king from being able to move there. Function not finished, just want to push the commit to be able to pull code on my laptop to work on. Ignore any current bugs due to this push, going to push an update with this fixed within the coming days